+++ CHANGELOG.md
@@ -4,6 +4,25 @@
 
 * Add support for bitwise operation in LEAN formula generation using Mathlib
 
+## v25.10.19
+
+### Bugfixes
+
+* Fix optimizer bug
+
+### Features
+
+* Support memory operations in formula generation
+* Support simple recursive functions in LEAN formula generation
+* Add Windows ARM64 build support
+
+### Enhancements
+
+* Add formula parser for parsing formulas from their string representation
+* Add internal command for finding programs with slow formula generation
+* Improve eval performance using optimized linker settings
+* Extend program stats with operation type bitmask
+
 ## v25.10.16
 
 ### Bugfixes